+++ src/main/java/com/example/team1_be/domain/Day/Day.java
@@ -16,7 +16,7 @@ import javax.validation.constraints.Size;
 public class Day {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Enumerated(EnumType.STRING)
     @NotNull
@@ -27,7 +27,7 @@ public class Day {
     private Week week;
 
     @Builder
-    public Day(int id, Weekday weekday, Week week) {
+    public Day(Long id, Weekday weekday, Week week) {
         this.id = id;
         this.weekday = weekday;
         this.week = week;
