+++ src/main/java/com/example/team1_be/domain/Notification/Notification.java
@@ -16,7 +16,7 @@ import javax.validation.constraints.Size;
 public class Notification {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 //
     @Column(length = 200, nullable = false)
     @Size(max = 200)
@@ -35,7 +35,7 @@ public class Notification {
     private User user;
 
     @Builder
-    public Notification(int id, String content, NotificationType type, Boolean isRead, User user) {
+    public Notification(Long id, String content, NotificationType type, Boolean isRead, User user) {
         this.id = id;
         this.content = content;
         this.type = type;
